+++ CHANGELOG.rst
@@ -4,4 +4,8 @@ Change Log
 Version 0.1
 -----------
 
-* TBD
+This is the initial release of the package and includes the following functionality:
+
+* Create a ROM/RC model using Modelica 3.2.x, Modelica Buildings Library 7.0 and TEASER 0.6.8
+* Create a Spawn-based models which loads an IDF file
+